eval_container/requirements.txt
mlflow==2.14.1
boto3==1.34.74
pandas==2.2.2
pyarrow==15.0.2
kazoo==2.10.0
kafka-python==2.0.2
python-dateutil==2.9.0
requests==2.32.3
fastapi==0.111.0
uvicorn==0.30.0
